First Code Commit
Get Ready to get mad! This commit adds some code that I worked on this
weakend. It gives us a basic frontier and also gives us a basic API
Fetcher. If you want to use it, just start the mockup.py in the
frontier folder and then start the tumblr_test.py in the APIFetch
Folder. This should all be run with python 3!